refactor: reorganize for easier running
`uvicorn main:app` should run now along with `pytest`
